% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(datalist, n)
}
\arguments{
\item{datalist}{Partial factor scores tables that we are sampling from}

\item{n}{Bootstrap sample number}
}
\value{
\describe{
    \item{t_value}{Bootstrap ratios, which, like t statistics,
    can be used to find the observations that reliably contribute
    to a given component.}
    \item{bt_mean}{Mean of the bootstrap samples}
    \item{bt_sd}{Standard deviation of the bootstrap samples}
    }
}
\description{
To estimate the stability of the compromise factor scores,
  we use bootstrap approach
}
\examples{
# Create an mfa object.
sets.num <- list(c(1:6), c(7:12), c(13:18), c(19:23), c(24:29), c(30:34),
  c(35:38), c(39:44), c(45:49), c(50:53))
mfa1 <- mfa(winedata, sets.num)
bootstrap(mfa1$Fpartial,1000)

}

