% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_table.R
\name{gentable}
\alias{gentable}
\title{General Table Function}
\usage{
gentable(data, sets, func = NA, dataname = NULL)
}
\arguments{
\item{data}{The data set whose between-table structure the user wishes
to study.}

\item{sets}{A list of vectors indicating how the \code{data} is organized
into sub-tables.}

\item{func}{A function that calculates a between-table coefficient. Must be
either \code{rv_coeff} or \code{lg_coeff}.}

\item{dataname}{An optional parameter. A character vector with the names
of the sub-tables. If the user supplies a value, \code{dataname} will be
used as both the rownames and colnames of the result matrix.}
}
\value{
The function returns a K by K matrix of between-table coefficients,
  where K is the number of sub-tables indicated by the \code{sets}
  parameter. The [i, j] element of the result matrix is the between-table
  coefficient between sub-table i and sub-table j.
}
\description{
Creates a table of Rv Coefficients or Lg Coefficients, based on
  the user's choice.
}
\examples{
# Create an mfa object.
sets.num <- list(c(1:6), c(7:12), c(13:18), c(19:23), c(24:29), c(30:34),
  c(35:38), c(39:44), c(45:49), c(50:53))
mfa1 <- mfa(winedata, sets.num)

# scaled data, Rv
gentable(mfa1$data_scale, as.list(c(sets.num[1], sets.char[2:3])),
          paste0("sub-table_", 1:3), func = rv_coeff)

# raw data, Rv
gentable(mfa1$origdata, as.list(c(sets.num[1], sets.char[2:3])),
          paste0("sub-table_", 1:3), func = rv_coeff)

# scaled data, Lg
gentable(mfa1$data_scale, as.list(c(sets.num[1], sets.char[2:3])),
          paste0("sub-table_", 1:3), func = lg_coeff)

# raw data, Lg
gentable(mfa1$origdata, as.list(c(sets.num[1], sets.char[2:3])),
          paste0("sub-table_", 1:3), func = lg_coeff)

}

