% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = "vector.norm")
}
\arguments{
\item{data}{A dataframe or matrix to be analyzed with MFA.}

\item{sets}{A list of numeric or character vectors indicating the blocks,
or sub-tables, into which the \code{data} is divided.}

\item{ncomps}{Integer indicating how many components (i.e. factors) are
to be extracted by the GSVD performed during MFA. \cr
By default, NULL means that the maximum number of components will be
  extracted (based on the rank of \code{data}).}

\item{center}{A logical value, or a numeric vector of length equal to the
number of active variables in the analysis. \cr
If center is a numeric vector, each variable (column) has the corresponding
  element subtracted from it. \cr
If TRUE, each column is mean-centered. \cr
If FALSE, no centering is done.}

\item{scale}{A logical value, a numeric vector of length equal to the
 number of active variables in the analysis, or the character string
 'vector.norm'. Scaling is performed after centering. \cr
 If TRUE, scaling is done by dividing each column by its
   standard deviation. \cr
If FALSE, no scaling is done. \cr
If scale is a numeric vector, each column is divided by the corresponding
  element of scale. \cr
  If scale is 'vector.norm' then each variable is divided by its own vector
    norm (Euclidean length).}
}
\value{
The function returns an S3 object of class 'mfa'. The object is a
  list with the following elements: \cr
 \describe{
   \item{a}{A vector of repeated table weights for the sub-tables.
     Each element of \code{alpha} is repeated J.k times, where J.k is
     the number of variables (columns) in the kth sub-table.}
   \item{alpha}{A vector of table weights. The kth element of \code{alpha}
      is equal to the first singular value of the kth subtable, raised to
      the power of -2. This is based on a standard PCA conducted on each
      sub-table, after centering and scaling.}

   \item{arglist}{A list of all the arguments the user passed to the
      \code{mfa} function.}

   \item{data_scale}{The original \code{data}, limited to the active
      columns (as determined by \code{sets}), after centering and scaling.}

   \item{dimdata}{The dimensions of \code{data_scale}.}

   \item{eigvals}{Eigenvalues from the GaSVD performed on the Grand Table.}

   \item{Fcommon}{A matrix of common/compromise factor scores.}

   \item{Fpartial}{A list of partial factor score matrices.}

   \item{gen_comps}{The number of components actually generated. \cr
      If the user chose an integer for \code{ncomps},
        then \code{gencomps} = \code{ncomps}. \cr
      Else if \code{ncomps} = NULL,
        then \code{gencomps} = \code{L}.}

   \item{K}{The number of sub-tables (blocks) into which the main data is
      divided, as determined by \code{sets}.}

   \item{L}{The rank of \code{data}.}

   \item{origdata}{A copy of the data frame or matrix the user passed to
     the \code{data} argument of the \code{mfa} function.}

   \item{Q}{A matrix of loadings/right singular vectors from the GSVD.}

   \item{sets}{A copy of the list passed to the \code{sets} argument.}

   \item{subtable}{A list of \code{K} matrices, where the kth element is
      the kth sub-table (block) of data, centered and scaled.}
 }
}
\description{
Creates an object of class \code{"mfa"}
}
\details{
This function computes the MFA using the 'MFA as Simple PCA' approach
   outlinedby Abdi, Williams, and Valentin on page 9 of "Multiple factor
   analysis: principal component analysis for multitable and multiblock data
   sets" (2013).

 STEP 1: Normalize each column of the original data. Typically this involves
   subtracting the mean and dividing by the vector norm. Split the data
   into sub-tables based on the different subjects (e.g. wine critics).

 Step 2: Perform standard PCA on each sub-table, and record each first
   singular value. Create the matrices of Row and Column constraints.
   Calculate GSVD using these constraints. This approach is laid out on
   Page 24 of "Principal Component Analysis" by Hevre Abdi and Lynne J
   Williams, published in Wiley Comp Stats, Volume 2, 2010.

 Step 3: Calculate quantities of interest relating to the GSVD. Put these
   in a nice list and return the list as an object of class 'mfa.'
}
\examples{
# Different ways to identify the sub-tables (blocks)
sets.num <- list(c(1:6), c(7:12), c(13:18), c(19:23), c(24:29), c(30:34),
  c(35:38), c(39:44), c(45:49), c(50:53))
sets.char <- list(c('V1','V2','V3','V4','V5','V6'),
  c('V1.1','V2.1','V3.1','V4.1','V7','V8'),
  c('V1.2','V2.2','V3.2','V4.2','V9','V10'),
  c('V1.3','V2.3','V3.3','V4.3','V8.1'),
  c('V1.4','V2.4','V3.4','V4.4','V11','V12'),
  c('V1.5','V2.5','V3.5','V4.5','V13'),
  c('V1.6','V2.6','V3.6','V4.6'),
  c('V1.7','V2.7','V3.7','V4.7','V14','V5.1'),
  c('V1.8','V2.8','V3.8','V4.8','V15'),
  c('V1.9','V2.9','V3.9','V4.9'))
sets.mixed <- c(sets.num[1:4], sets.char[5:10])

# default values for ncomps, center, and scale
mfa1 <- mfa(winedata, sets.mixed)

# Only generate the first 2 components
mfa2 <- mfa(winedata, sets.mixed, ncomps = 2)

# Cannot ask for more components than the rank of data
\dontrun{
mfa(winedata, sets.mixed, ncomps = 22)
}

Cannot have elements in sets that are not column names/indices
\dontrun{
mfa(winedata, c(sets.mixed, "AlfredRenyi"))
mfa(winedata, c(sets.mixed, 18180505))
}

If center or scale are vectors, they must be equal in
  length to the number of active columns (variables).
mfa3 <- mfa(winedata, sets.mixed, center = 1:53, scale = 1:53)
\dontrun{
mfa3 <- mfa(winedata, sets.mixed, scale = 1:52)
mfa3 <- mfa(winedata, sets.mixed, center = 1:54)
}

}

