% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mfa}
\alias{plot.mfa}
\title{Plot Method For "mfa" Object}
\usage{
\method{plot}{mfa}(x, type, xdim = 1, ydim = 2, facetrows = 2, size = 5,
  subtabs = NULL, legend = NA, mytitle = NA, label = NA,
  bootstrap_size = 1000, bootstrap_comps = c(1, 2))
}
\arguments{
\item{x}{An object of class "mfa".}

\item{type}{Indicates what type of plot the user wishes to see. Must be one
of the following character strings: "compromise", "partial.factor", or
"loadings".}

\item{xdim, ydim}{The two components the user wishes to plot. Numeric values.}

\item{facetrows}{Used with "partial.factor" and "loadings" plots. Controls
how many rows to use when displaying multiple sub-plots.}

\item{size}{Controls the size of the plotted points. If plotted points
overlap, the user is encouraged to try reducing size.}

\item{subtabs}{Used with "partial.factor" and "loadings" plots. Allows the
user to choose which sub-tables she/he wants to see plots for. Default is
NULL, which will display all the subtables. If not NULL, must be a numeric
vector. each element must be between 1 and K, where K is the total number
of sub-tables in the analysis.}

\item{legend}{An optional parameter that allows the user to control legend
text. Default value is NA. If NA, legend text will be chosen
automatically, based on data row or column names, depending on the plot
type.}

\item{mytitle}{An optional parameter for the user to choose the plot title.
By default, mytitle is NA. If NA, the plot is given a title corresponding
to its type, viz "compromise", "partial.factor", or "loadings."}

\item{label}{Used with "compromise" and "compromise.partial" plots. Allows the
user to choose which values can be presented as a label on the plot. Default is
NULL, which will display no label. If not NULL, must be a vector.}

\item{bootstrap_size}{Used only with "bootstrap" plot to control the bootstrap size. Default
value is 1000.}

\item{bootstrap_comps}{Used only with "bootstrap" plot. Allows the user to chosose which
components of bootstrap result she/he want to see plots for. Default is c(1,2), which will
display component 1 and 2.}
}
\value{
Displays the plot of the user's choice.
}
\description{
A plotting function that, given two components/dimensiosn (e.g.
  1 & 2), displays a graphic of one of the following: \cr
  \itemize{
    \item Compromise/Common Factor Scores
    \item Partial Factor Scores
    \item Loadings
    \item Eigenvalues
    \item Compromise + Partial Factor Scores
    \item Bootstrap ratio plots
  }
}
\examples{
# Create an mfa object.
sets.num <- list(c(1:6), c(7:12), c(13:18), c(19:23), c(24:29), c(30:34),
  c(35:38), c(39:44), c(45:49), c(50:53))
mfa1 <- mfa(winedata, sets.num)

# Different types of plots:
plot(mfa1, type = "compromise", legend=substr(rownames(mfa1$Fcommon),1,2),
label=substr(rownames(mfa1$Fcommon),3,3))
plot(mfa1, type = "partial.factor",  subtabs = NULL, xdim = 2, ydim = 3, size = 4,
legend=substr(rownames(mfa1$Fpartial[[1]]),1,2), label=substr(rownames(mfa1$Fcommon),3,3))
plot(mfa1, type = "loadings", size = 2.5, subtabs = c(9,10),
legend = c("cat pee", "passion fruit", "green pepper", "mineral","optional 1", "optional 2")))
plot(mfa1, type = "eigenvalues")
plot(mfa1, type = "compromise.partial", xdim = 1, ydim = 2, l
egend=substr(rownames(mfa1$Fcommon),1,2),label=substr(rownames(mfa1$Fcommon),3,3))
plot(mfa1, type= "bootstrap", bootstrap_size = 1000, bootstrap_comps=c(1,2), facetrows=2)

}

