% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_table.R
\name{rv_coeff}
\alias{rv_coeff}
\title{Rv Coefficient}
\usage{
rv_coeff(X, Y)
}
\arguments{
\item{X, Y}{Two data tables (matrices or data frames)}
}
\value{
The function returns the Rv Coefficient, a numeric vector, length 1.
  This is a multivariate generalization of the squared Pearson Product-Moment
  Correlation Coefficient. It takes real values from 0 to 1.
}
\description{
Calculates the Rv Coefficient between two data tables. This is
  a measure of similarity between data tables.
}
\details{
Calculated according to Equation (29), page 8 of 'Abdi, Herve,
  Lynne J. Williams, and Domininique Valentin. "Multiple Factor Analysis:
  PrincipalComponent Analysis for Multitable and Multiblock Data Sets." Wiley
  Interdisciplinary Reviews: Computational Statistics 5.2 (2013): 149-79'
}
\examples{
# Create an mfa object.
sets.num <- list(c(1:6), c(7:12), c(13:18), c(19:23), c(24:29), c(30:34),
  c(35:38), c(39:44), c(45:49), c(50:53))
mfa1 <- mfa(winedata, sets.num)

# wine data: sub-tables 1 and 2, normalized
rv_coeff(mfa1$subtable[[1]], mfa1$subtable[[2]])

# wine data: sub-tables 1 and 2, non-normalized
rv_coeff(winedata[,1:6], winedata[,7:12])

}

