% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{varcont}
\alias{varcont}
\title{Variable Contribution Method}
\usage{
varcont(x, verbose = FALSE)
}
\arguments{
\item{x}{An object of class "mfa".}

\item{verbose}{A flag that controls whether the summary table is printed
to the console. \cr
If TRUE, then the summary table is printed when the function is called. \cr
If FALSE, no output is printed. \cr
In either event, the summary table may be assigned to another object and
  accessed later.}
}
\value{
The function returns a matrix of variable contributions. The
  [i, j] element of this matrix (ith row, jth column) is the contribution
  of the ith variable to the jth component.
}
\description{
Creates a table that shows the contribution of each variable
  to each component.
}
\details{
Calculated according to Equation (27), page 7 of 'Abdi, Herve,
  Lynne J. Williams, and Domininique Valentin. "Multiple Factor Analysis:
  PrincipalComponent Analysis for Multitable and Multiblock Data Sets." Wiley
  Interdisciplinary Reviews: Computational Statistics 5.2 (2013): 149-79'
}
\examples{
# Create an mfa object.
sets.num <- list(c(1:6), c(7:12), c(13:18), c(19:23), c(24:29), c(30:34),
  c(35:38), c(39:44), c(45:49), c(50:53))
mfa1 <- mfa(winedata, sets.num)
## Two ways to use this function
# 1. Print the output immediately
varcont(mfa1, verbose = TRUE)
# 2. silently assign the object to another object without displaying
summary <- varcont(mfa1)
round(summary, 3)[1:5,1:5]

}

